LayoutTests/http/tests/loading/progress-finished-callback.html

<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpProgressFinishedCallback();
}
</script>
This is a test of progress finished callback ordering in relation to frame loader callbacks. It is only useful inside the regression test tool.
